% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_comp.R
\name{bc_comp}
\alias{bc_comp}
\title{Calculates Percentage of core compression and Linear Correction Factor}
\usage{
bc_comp(tube_lenght, core_in, core_out)
}
\arguments{
\item{tube_lenght}{The lenght in cm of the sampler.}

\item{core_in}{The lenght in cm of the part of the sampler left outside of the sediment (from the inside of the sampler).}

\item{core_out}{The lenght in cm of the part of the sampler left outside of the sediment (from the outside of the sampler).}
}
\value{
Percentage of core compression and the Linear Correction Factor.
}
\description{
This function calculates the Percentage of core compression and the Linear Correction Factor using three arguments.
}
\examples{
bc_comp(200, 48, 25)
}
