% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_data.R
\docType{data}
\name{bc_cores}
\alias{bc_cores}
\title{Blue Carbon sediment core data}
\format{
Two dataframes, with core and sample properties
\describe{
\item{bc_cores}{TODO}
\item{bc_samples}{TODO}
}
}
\usage{
data(bc_data)
}
\description{
Example data from cores collected for Blue Carbon
stock estimations in Portugal. Cores have been de-identified.
}
\examples{
data(bc_data)
TODO
}
\keyword{datasets}
