% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlueCarbon-package.R
\docType{package}
\name{BlueCarbon-package}
\alias{BlueCarbon}
\alias{BlueCarbon-package}
\title{BlueCarbon: Estimation of Organic Carbon Stocks and Fluxes From Cores Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/BlueCarbon}
  \item Report bugs at \url{https://github.com/EcologyR/BlueCarbon/issues}
}

}
\author{
\strong{Maintainer}: Nerea Piñeiro-Juncal \email{npineiro@ua.pt} (\href{https://orcid.org/xxxx}{ORCID})

Authors:
\itemize{
  \item Julen Astigarraga
}

}
\keyword{internal}
