% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_comp.R
\name{calculate_core_compaction}
\alias{calculate_core_compaction}
\title{Estimate compaction rate for cores
Accepts a data.frame with core properties and returns a modified version
of it, with the addition of the estimated parameters}
\usage{
calculate_core_compaction(
  data,
  tube_length,
  internal_distance,
  external_distance
)
}
\arguments{
\item{tube_length:}{name for column with length of the sampling tube used}

\item{internal_distance:}{name for column with distance between sampler top and the sediment core surface}

\item{external_distance:}{name for column with distance between sampler top and the sediment column surface}
}
\description{
Estimate compaction rate for cores
Accepts a data.frame with core properties and returns a modified version
of it, with the addition of the estimated parameters
}
