% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompact_linear.R
\name{decompact_linear}
\alias{decompact_linear}
\title{Calculate sediment properties after decompression}
\usage{
decompact_linear(
  df = NULL,
  df_fm = NULL,
  core = "core",
  compression = "compression",
  mind = "mind",
  maxd = "maxd"
)
}
\arguments{
\item{df}{data.frame with core properties}

\item{df_fm}{data.frame with field measurements, used to calculate compression
if not provided}

\item{compression}{name of the column with core compression IN PERCENTAGE}

\item{mind}{name of the column with minimum depth}

\item{maxd}{name of the column with maximum depth}
}
\value{
None
}
\description{
Accepts a data.frame with sample properties and returns a modified version
with sample properties after corrected for compression
}
\examples{
plot_crayons()

}
