% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_oc.R
\name{estimate_oc}
\alias{estimate_oc}
\title{Estimate organic carbon}
\usage{
estimate_oc(
  df = NULL,
  core = "core",
  site = "site",
  ecosystem = "ecosystem",
  species = "species",
  om = "om",
  oc = "oc"
)
}
\arguments{
\item{df}{A tibble or data.frame containing all the data. Must have at least
five columns (see arguments below).}

\item{core}{TODO}

\item{site}{Character Name of the column reporting sample site.}

\item{ecosystem}{Character Name of the column reporting ecosystem type.
To apply published equations for OC estimation, ecosystem names should be
either "Salt Marsh", "Seagrass" or "Mangrove".}

\item{species}{Character Name of the column reporting the main species in the site.}

\item{om}{Character Name of the column reporting organic matter values.}

\item{oc}{Character Name of the column reporting organic carbon values.}
}
\value{
The initial tibble or data.frame with three new columns:
\itemize{
\item one column with estimated organic carbon values (eOC) in \%
\item the standard error of the prediction (eOC_se)
\item the type of model used for estimation (origin)
In addition, a plot with the relationship between organic matter and estimated
organic carbon
}
}
\description{
Estimate organic carbon from organic matter values
}
\details{
Estimation of organic Carbon is done by means of linear regressions on
log(organic carbon) ~ log(organic matter). It gives back a organic carbon
value for each organic matter value provided. If there is a organic carbon
value for that sample it return the same value, else, generates a model for
that site, else, model for specie, else,  model for Ecosystem. If a model
can not be created due to the low number of samples (<10) it uses the equations
in Howard et al. 2014 to estimate the organic carbon.
}
\examples{
bluecarbon_decompact <- decompact(bluecarbon_data)
out <- estimate_oc(bluecarbon_decompact)
}
