% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_stock.R
\name{estimate_stock}
\alias{estimate_stock}
\title{Organic Carbon Stock estimation}
\usage{
estimate_stock(df = NULL, Depth = 100)
}
\arguments{
\item{df}{A \link{data.frame} with, at least, columns: Core.ID, Min.D (minimum depth of the sample), Max.D (maximum depth of the sample), DBD (dry bulk density), POC (organic carbon \%)}

\item{Depth}{standardization soil depth, by default 100 cm.}
}
\value{
\link{data.frame} with columns Core.id, S.WC (organic carbon stock at the whole core), D.Max (maximum depth of the core), and Stock (organic carbon stock at the standardized depth)
}
\description{
Estimates the carbon stock from soil core data until a specific depth, 100 cm by default. If the core do not reach the
standardization depth it extrapolate the stock from a linear model between the organic carbon accumulated mass and depth.
}
