% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_diagnostic.R
\name{plot_stock_diagnostic}
\alias{plot_stock_diagnostic}
\title{plot_stock_diagnostic}
\usage{
plot_stock_diagnostic(
  sample_data,
  core_id,
  sample_depth,
  element_concentration,
  maximum_depth,
  method
)
}
\arguments{
\item{sample_data}{Sample data as a list, split by sediment core}

\item{method}{Method used to calculate elemental stock}
}
\description{
Internal function used to plot stock estimations in \code{bc_stocks}
}
