% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_extrapolation.R
\name{test_extrapolation}
\alias{test_extrapolation}
\title{Test differences between observed and extrapolated stocks}
\usage{
test_extrapolation(df = NULL, Depth = 100)
}
\arguments{
\item{df}{A \link{data.frame} with, at least, columns: Core.ID, Min.D (minimum depth of the sample), Max.D (maximum depth of the sample), DBD (dry bulk density), POC (organic carbon \%)}

\item{Depth}{standardization soil depth, by default 100 cm.}
}
\value{
A \link{data.frame} with the observed and extrapolated stocks. A plot with comparisons.
}
\description{
subset those cores that reach the standardization depth and estimates the stock (observed stock), estimate the stock from the linear relation of organic carbon accumulated mass and depth using the 90, 75, 50 and 25\% top length of the indicated standardization depth. Compares the observed stock with the estimated stocks by extrapolation.
}
