% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_om_oc.R
\name{transform_om_oc}
\alias{transform_om_oc}
\title{Organic carbon estimation}
\usage{
transform_om_oc(df = NULL)
}
\arguments{
\item{df}{A \link{data.frame} with, at least, columns Core.ID, Ecosystem, Specie, Site.ID, OM, and OC.}
}
\value{
the initial \link{data.frame} + one column with organic carbon values
}
\description{
Model linear relation between organic matter and organic carbon and estimate organic carbon values from organic matter data
}
