% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_compaction.R
\name{bc_compaction}
\alias{bc_compaction}
\title{bc_compaction}
\usage{
bc_compaction(
  core_data,
  sampler_length = "sampler_length",
  internal_distance = "internal_distance",
  external_distance = "external_distance"
)
}
\arguments{
\item{core_data}{data.frame with core properties}

\item{sampler_length}{name of the column with the total length of the sampler tube}

\item{internal_distance}{name of the column with distance between sampler top and core surface}

\item{external_distance}{name of the column with distance between sampler top and sediment surface}
}
\value{
the initial data.frame with the addition of Percentage of core compression
}
\description{
Calculates Percentage of core compression for cores
Accepts a data.frame with core properties and returns a modified version
of it, with the addition of the estimated parameters
}
