% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_decomp.R
\name{bc_decomp}
\alias{bc_decomp}
\title{bc_decomp}
\usage{
bc_decomp(data, tube_lenght, core_in, core_out, diameter, method = "linear")
}
\arguments{
\item{data}{dataframe with the following columns "ID" "cm" "weight" "LOI" "c_org".}

\item{tube_lenght}{The lenght in cm of the sampler.}

\item{core_in}{The lenght in cm of the part of the sampler left outside of the sediment (from the inside of the sampler).}

\item{core_out}{The lenght in cm of the part of the sampler left outside of the sediment (from the outside of the sampler).}

\item{diameter}{in cm of the sampler}

\item{method}{used to estimate the decompressed depth of each section, "linear" or "exp". Default is "linear".}
}
\value{

}
\description{
This function uses six arguments
}
\examples{
bc_decomp(data, 200, 48, 25, 6.9, method = "linear")
}
