% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_stocks.R
\name{bc_stocks}
\alias{bc_stocks}
\title{estimate_stock()}
\usage{
bc_stocks(
  sample_data,
  core_id,
  sample_depth,
  element_concentration,
  maximum_depth,
  method = NULL,
  section_height = NULL,
  diagnostic_plot = FALSE
)
}
\arguments{
\item{sample_data}{}
}
\description{
Estimate the stock of an element contained given its concentration along
a sediment core
}
\details{
\strong{Calculate elemental stocks:}
\enumerate{
\item Trapezoid rule (Martins, M. et al 2021)
\item Rectangle rule (Howard, J. et al 2014)
}
\itemize{
\item User provides height of sections represented by samples
\item Height of section represented by samples are automatically estimated
3.(optional) - Plots how the stock is being estimated, with a special draw towards any sections that have to be extrapolated
}
}
