% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_stocks.R
\name{bc_stocks}
\alias{bc_stocks}
\title{estimate_stock()}
\usage{
bc_stocks(
  sample_data,
  core_id,
  sample_depth,
  element_concentration,
  maximum_depth,
  method = NULL,
  section_height = NULL,
  diagnostic_plot = FALSE
)
}
\arguments{
\item{sample_data}{}
}
\description{
Estimate the stock of an element contained given its concentration along
a sediment core
}
