% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlueCarbon-package.R
\docType{package}
\name{BlueCarbon-package}
\alias{BlueCarbon}
\alias{BlueCarbon-package}
\title{BlueCarbon: Estimation of Organic Carbon Stocks and Fluxes From Soil Core Data}
\description{
BlueCarbon package intends to provide tools to estimate soil organic carbon stocks and fluxes in blue carbon ecosystemses. It contains seven main functions to estimate the compresion of cores, matematially correct core comprasion, estimate sample tchikness, estimate organic carbon content from organic matter content, estimate organic carbon stocks fluxes and visualize the error of stock extrapolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/BlueCarbon}
  \item Report bugs at \url{https://github.com/EcologyR/BlueCarbon/issues}
}

}
\author{
\strong{Maintainer}: Nerea Piñeiro-Juncal \email{npineiro@ua.pt} (\href{https://orcid.org/xxxx}{ORCID})

Authors:
\itemize{
  \item Julen Astigarraga
}

}
\keyword{internal}
