% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlueCarbon-package.R
\docType{package}
\name{BlueCarbon-package}
\alias{BlueCarbon}
\alias{BlueCarbon-package}
\title{BlueCarbon: Estimation of Organic Carbon Stocks and Sequestration Rates from Soil Core Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to estimate soil organic carbon stocks and sequestration rates in blue carbon ecosystems. 'BlueCarbon' contains functions to estimate and correct for core compaction, estimate sample thickness, estimate organic carbon content from organic matter content, estimate organic carbon stocks and sequestration rates, and visualize the error of carbon stock extrapolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/BlueCarbon}
  \item \url{https://ecologyr.github.io/BlueCarbon/}
  \item Report bugs at \url{https://github.com/EcologyR/BlueCarbon/issues}
}

}
\author{
\strong{Maintainer}: Nerea Piñeiro-Juncal \email{np.juncal@gmail.com} (\href{https://orcid.org/0000-0003-3767-1812}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Julen Astigarraga (\href{https://orcid.org/0000-0001-9520-3713}{ORCID})
  \item Valentina Costa (\href{https://orcid.org/0000-0002-1513-0284}{ORCID})
  \item Marcio Martins (\href{https://orcid.org/0000-0002-6969-2215}{ORCID})
  \item Francisco Rodriguez-Sanchez (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
}

}
\keyword{internal}
