% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompact.R
\name{decompact}
\alias{decompact}
\title{Calculate sediment properties after decompaction}
\usage{
decompact(
  df = NULL,
  core = "core",
  compaction = "compaction",
  mind = "mind",
  maxd = "maxd",
  dbd = NULL
)
}
\arguments{
\item{df}{data.frame with core properties}

\item{compaction}{name of the column with core compaction IN PERCENTAGE,
as calculated with \code{\link[=estimate_compaction]{estimate_compaction()}}.}

\item{mind}{name of the column with minimum depth}

\item{maxd}{name of the column with maximum depth}

\item{dbd}{name of the column with dry bulk density}
}
\value{
the initial data.frame with the addition of two columns with the corrected minimum and maximum depth of the samples
(additionally, if a dry bulk density column is specified, it will return another column with corrected dry bulk density)
}
\description{
Accepts a data.frame with sample properties and returns a modified version
with sample properties after corrected for compaction
}
\examples{
decompact(bluecarbon_data)

}
