% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_compaction.R
\name{estimate_compaction}
\alias{estimate_compaction}
\title{Estimate core compaction}
\usage{
estimate_compaction(
  df,
  core = "core",
  sampler_length = "sampler_length",
  internal_distance = "internal_distance",
  external_distance = "external_distance"
)
}
\arguments{
\item{df}{data.frame with core properties}

\item{sampler_length}{name of the column with the total length of the sampler tube}

\item{internal_distance}{name of the column with distance between sampler top and core surface}

\item{external_distance}{name of the column with distance between sampler top and sediment surface}

\item{core_id}{core identification}
}
\value{
the initial data.frame with the addition of Percentage of core compression
}
\description{
Calculates Percentage of core compression for cores
Accepts a data.frame with core field measurements and returns a modified version
of it, with the addition of the estimated parameters
}
\examples{
TODO

}
