% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_flux.R
\name{estimate_flux}
\alias{estimate_flux}
\title{Estimate the average organic carbon flux}
\usage{
estimate_flux(
  df = NULL,
  timeframe = 100,
  core = "core",
  mind = "mind_corrected",
  maxd = "maxd_corrected",
  dbd = "dbd",
  oc = "eoc",
  age = "age"
)
}
\arguments{
\item{df}{A data.frame with, at least, columns: core, mind (minimum depth of the sample),
maxd (maximum depth of the sample), dbd (dry bulk density), oc (organic carbon \%),
age (age of the sample obtained from a age-depth or age-accumulated mass model)}

\item{timeframe}{standardization time frame, by default 100 years}

\item{core}{Character Name of the column reporting core ID.}

\item{mind}{Character Name of the column reporting the minimum depth of each sample.}

\item{maxd}{Character Name of the column reporting the maximum depth of each sample.}

\item{dbd}{Character Name of the column reporting dry bulk density.}

\item{oc}{Character Name of the column reporting organic carbon concentrations.}

\item{age}{Character Name of the column reporting the age of each sample.}
}
\value{
data.frame with columns Core.id, F.WC (organic carbon fluxes at the whole core),
A.Max (maximum age of the core), and Flux (average organic carbon flux at the indicated time frame)
}
\description{
estimate the average organic carbon flux to the soil in a indicated time frame (by default last 100 years)
from the organic carbon concentration and ages obtained from a age-depth or age-accumulated mass model
}
\examples{
bluecarbon_decompact <- decompact(bluecarbon_data)
oc <- estimate_oc(bluecarbon_decompact)
out <- estimate_flux(oc)
}
