% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_flux.R
\name{estimate_flux}
\alias{estimate_flux}
\title{Estimate the average organic carbon flux}
\usage{
estimate_flux(df = NULL, TimeFrame = 100)
}
\arguments{
\item{df}{A \link{data.frame} with, at least, columns: Core.ID, Min.D (minimum depth of the sample), Max.D (maximum depth of the sample), DBD (dry bulk density), POC (organic carbon \%), Age (age of the sample obtained from a age-depth or age-accumulated mass model)}

\item{TimeFrame}{standardization time frame, by default 100 years}
}
\value{
\link{data.frame} with columns Core.id, F.WC (organic carbon fluxes at the whole core), A.Max (maximum age of the core), and Flux (average organic carbon flux at the indicated time frame)
}
\description{
estimate the average organic carbon flux to the soil in a indicated time frame (by default last 100 years) from the organic carbon concentration and ages obtained from a age-depth or age-accumulated mass model
}
