% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_h.R
\name{estimate_h}
\alias{estimate_h}
\title{Estimation of the thickness of the sample}
\usage{
estimate_h(df = NULL, core = "core", mind = "mind", maxd = "maxd")
}
\arguments{
\item{df}{A \link{data.frame} with with columns core, mind (minimum depth of the sample)and maxd (maximum depth of the sample)}

\item{core}{Character Name of the column reporting core ID.}

\item{mind}{Character Name of the column reporting the minimum depth of each sample.}

\item{maxd}{Character Name of the column reporting the maximum depth of each sample.}
}
\value{
the initial \link{data.frame} with three additional columns: emin (estimated minimum depth of the sample), emax (estimated maximum depth of the sample) and h (estimated thikness of the sample)
}
\description{
checks for space between samples and, if any, divide this space between the previous and
next sample to return sample thickness withouth gaps in the core
}
\examples{
estimate_h(A)
}
